feat!: change customPrettifiers to a Map
In [1], LevelPrettifierExtras and PrettifierExtras were merged to solve
a typing issue [2]. This was suboptimal as the additional extra
attributes that only the 'level' prettifier had were exposed to other
prettifiers too.

To allow the additional attributes while keeping 'level' prettifier's
type separate, this patch uses an intersection Map type
CustomPrettifiers instead.

This is a breaking change.

[1] https://github.com/pinojs/pino-pretty/pull/551
[2] https://github.com/pinojs/pino-pretty/issues/550